+++ src/index.d.ts
@@ -1590,6 +1590,18 @@ declare namespace connect {
     /** Determine whether the connection is mute server side. */
     isMute(): boolean;
 
+    /** Start contact recording */
+    startContactRecording(): Promise<any>;
+
+    /** Stop contact recording */
+    stopContactRecording(): Promise<any>;
+
+    /** Resume contact recording */
+    suspendContactRecording(): Promise<any>;
+
+    /** Suspend contact recording */
+    resumeContactRecording(): Promise<any>;
+
     /**
      * Mute the connection server side
      *